fix: revert the error no the lexer kinda work again
TODO: find a way to deal with float numbers

+++ include/Expr/Lexer/Value.hpp
@@ -4,7 +4,7 @@
 #include <variant>
 
 // TokenValue
-using value_t = std::variant<std::monostate, int, std::string>;
+using value_t = std::variant<std::monostate, double, std::string>;
 
 class Value
 {
